# backend/tests/openapi.json
{
 "components": {
  "schemas": {
...
   "ChatRequest": {
    "additionalProperties": false,
    "properties": {
     "chip_id": {
      "anyOf": [
       {
        "enum": [
         "today",
         "why",
         "history",
         "after_crisis",
         "vix"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "A suggestion chip; its fixed question is used",
      "title": "Chip Id"
     },
     "message": {
      "maxLength": 2000,
      "minLength": 1,
      "title": "Message",
      "type": "string"
     },
     "session_id": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "Omit to start a new conversation",
      "title": "Session Id"
     }
    },
    "required": [
     "message"
    ],
    "title": "ChatRequest",
    "type": "object"
   },
...
   "ErrorBody": {
    "properties": {
     "error": {
      "$ref": "#/components/schemas/ErrorDetail"
     }
    },
    "required": [
     "error"
    ],
    "title": "ErrorBody",
    "type": "object"
   },
   "ErrorDetail": {
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "message": {
      "title": "Message",
      "type": "string"
     },
     "request_id": {
      "title": "Request Id",
      "type": "string"
     }
    },
    "required": [
     "code",
     "message",
     "request_id"
    ],
    "title": "ErrorDetail",
    "type": "object"
   },
   "FeedbackRequest": {
    "additionalProperties": false,
    "properties": {
     "message_id": {
      "description": "The message_id from the chat's done event",
      "minimum": 1.0,
      "title": "Message Id",
      "type": "integer"
     },
     "rating": {
      "enum": [
       "up",
       "down"
      ],
      "title": "Rating",
      "type": "string"
     },
     "session_id": {
      "description": "The session_id from the chat's done event",
      "format": "uuid",
      "title": "Session Id",
      "type": "string"
     }
    },
    "required": [
     "session_id",
     "message_id",
     "rating"
    ],
    "title": "FeedbackRequest",
    "type": "object"
   },
...
  }
 },
 "info": {
  "description": "Market-regime data and an educational AI analyst. Not investment advice.",
  "title": "MarketMood API",
  "version": "0.2.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/v1/chat": {
   "post": {
    "operationId": "chat_v1_chat_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChatRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       },
       "text/event-stream": {}
      },
      "description": "SSE stream (see module docs)"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unprocessable Entity"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Service Unavailable"
     }
    },
    "summary": "Chat",
    "tags": [
     "chat"
    ]
   }
  },
  "/v1/feedback": {
   "post": {
    "description": "Thumbs up / down on one AI answer of your own conversation (CHAT-09); the last click wins.",
    "operationId": "feedback_v1_feedback_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FeedbackRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unprocessable Entity"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Service Unavailable"
     }
    },
    "summary": "Feedback",
    "tags": [
     "chat"
    ]
   }
  },
...
 }
}